Logging improvements and C++23
